LinkedList dependency and modified auxiliary classes.
I used the library "linked-list" because I did not trust my own implementation. However, this implementation wasn't easy to work with. It requires to create new `Item (Node)` objects and modify its `.value` every time you want to add a new node.

The auxiliary classes `ModifiedSet` and `ModifiedMap` were necessary just because for ease of retrieval of keys. You can't use arrays as keys in a Map and also the `.has()` from Set. So I needed to parse the arrays as strings everytime I was about to retrieve, set, or check for some value. To avoid clogging the code with the parsing process in every part that I need to use `.get`, `.set` or `.has` methods, I made modified methods of these. In order to do this, It was needed to create a new class or modify the prototype of the orignal classes. I prefered the former.