[FIX]: Neighbors (adjacents) addition to the queue in the BFS algorithm
This is the origin of the bug. I was adding the neighbors of the first neighbor recursively before going to the other neighbors in the original node.

For example, the neighbors of [0,0] are [[1,2], [2,1]]. Instead of adding to the queue these two neighbors and then adding the neighbors of these two neighbors and so on (recursively), I was adding [1,2] and then the neighbors of [1,2] and then the neighbors of the neighbors of [1,2].

This was defeating the Breadth First Search algorithm and instead was a Depth-First-Search algorithm. This is why I was not getting the shortest path. I fixed the algorithm to add to the que all neighbors before going to the next level of depth.

After this, I had a good BFS algorithm, but the path tracking was not ready yet.