[FIX] Path tracking
The path tracking was also stuck in a Depth First fashion instead of a Breadth First. In order to fix it, I moved the part that tracks the path and adds it to the queue. I moved it to the inner while loop, this way the previous path would be updated with the node that follows in the Bread First algorith.